example_random_art/pyproject.toml
[tool.poetry]
name = "generative_art"
version = "1.0.0"
description = "Random Art Generator"
authors = ["Eric Davidson", "Michael Wittmann <michael.wittmann@tum.de"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.0.1"
image = "^1.5.33"
flake8 = "^3.8.4"
pytest = "^6.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
